Extracts common JsonMapper/TypeResolver factories
