+++ dataformat-json/src/main/java/org/eclipse/digitaltwin/aas4j/v3/dataformat/json/SimpleAbstractTypeResolverFactory.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2023 Fraunhofer-Gesellschaft zur Foerderung der angewandten Forschung e. V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.eclipse.digitaltwin.aas4j.v3.dataformat.json;
+
+import org.eclipse.digitaltwin.aas4j.v3.dataformat.core.util.ReflectionHelper;
+
+import com.fasterxml.jackson.databind.module.SimpleAbstractTypeResolver;
+
+/**
+ * Factory for creating a {@link SimpleAbstractTypeResolver} configured with the
+ * Default Implementations of AAS4J
+ * 
+ * @author schnicke
+ *
+ */
+public class SimpleAbstractTypeResolverFactory {
+    @SuppressWarnings("unchecked")
+    public SimpleAbstractTypeResolver create() {
+        SimpleAbstractTypeResolver typeResolver = new SimpleAbstractTypeResolver();
+        ReflectionHelper.DEFAULT_IMPLEMENTATIONS.stream().forEach(x -> typeResolver.addMapping(x.getInterfaceType(), x.getImplementationType()));
+        return typeResolver;
+    }
+}
